backend/requirements.txt
fastapi
uvicorn
pillow
google-genai
python-multipart
python-dotenv